pyproject.toml
[tool.poetry]
name = "makerspaceleiden-crm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.0.0"
pytz = "^2023.3.post1"
wheel = "^0.42.0"
django-import-export = "^3.3.3"
django-simple-history = "^3.4.0"
qrcode = "^7.4.2"
django-settings-export = "^1.2.1"
django-stdimage = "^6.0.2"
django-dynamic-filenames = "^1.3.2"
django-six = "^1.0.5"
six = "^1.16.0"
django-revproxy = "^0.12.0"
pillow = "^10.1.0"
django-qr-code = "^3.1.1"
django-ipware = "^6.0.0"
reportlab = "^4.0.7"
pypdf2 = "^3.0.1"
lxml = "^4.9.3"
jsonfield2 = "^4.0.0.post0"
django-money = "^3.4.0"
requests = "^2.31.0"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
pre-commit = "^3.5.0"

[tool.black]
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
ignore = ["E501", "E741"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
